fix(wrapper): add mjs wrapper to support esm
rescript-jotai is configured in commonjs. When using the package in a rescript project using es6, running "rescript build -with-deps" resulted in an error.
This was caused by the wrapper JS-file that is written in commonjs. By adding an alternative version in esm (wrapper.mjs), the correct version will be referenced automatically (by most build tools).